Replace symbol layer colon with asterisk
There were some complication around having the semicolon as a hold on both base and symbol layer. As it stands I will just have to use Shift and base layer semicolon to get a colon, which is not ideal as that will be the only symbol that is true for, lets see how it goes.