Keep the middle three keys of the thumb clusters
I tried going the middle three, like it would be with the Aysu, but it didnt work out, and honestly what am I trying to do there? Ive already removed the sixth column, the two upper thumb keys, and am replacing the outer one with the a knob. If I ever did have to make the change it will be fine, but until then, this is good. Am doubling up on the enters, but will try to move towards the centre of the cluster.