Merge num layer with nav layer
Might take a little getting used to having the nums on the left hand, but having the num layer mod not on the thumbs was not working well. Nums are used often enough that they warrant a home-thumb mod.

This also necessitated removing the ability to get to RGB layer with both thumbs, as now the left thumb needs to be a 0 in the numpad.