Reduce the max RGB value to 240
I have been having issues when the brightness is higher than this during
Rainbow swirl on my light Kyria. It is something to do with doubling up
the LEDs with the extra strip, perhaps drawing too much power?

Anyway it's not a noticeable difference from the max 255, this is fine.