Create Numbers layer on right hand
I just cannot get used the numpad on the left hand, and there is no way
directionals are moving from the vim positions, so a new layer is
essential.

I have decided to make assign the layer to the outer thumb, and move the
auxiliary Cmd key to the inner thumb. This actually puts it pretty much
the same position that it would be on a standard Mac keyboard, so
hopefully that should tap into some old muscle memory.

I am not re-using the old num positions, so as usual for this kind of
change I will leave them there as I get used to the new, and remove them
in time.