Reorder home row mod-taps, move shift to index finger
I don't know how easy this will be to get used to, but there I've seen
enough recommendations out there for putting Shift on the index finger
that I'm going to give it a go.

Specifically the Hands down layout says it is designed with that
expectation and I want to try that at some point.

Also it will make the most common mod combo of Cmd+Shift a better chord,
and I am hoping that it will help with the amount of misfires I get with
shifting during normal text typing, e.g getting 'ai' instead of 'I'.