Make the right home thumb a one shot shift
Move tab to outer thumb, need that on base layer for Cmd-tabbing. Del
was already on the Nav layer so it could be held, so no need to do
anything with that one.

It is not possible to have combine a one-shot with a layer change, so
it's all done manually using tap-dance, with special cases for handling
the home row mod-taps. Mostly works fine, although if you follow the
one-shot with mod-tap hold it will kill the mode. As the one-shot should
only be followed by a letter I can live with this.