Add one shot shift to right thumb cluster too
Also moved the left one, so that it is in the same position for each
hand.

Moved Enter to the outer most thumb key. I have been aiming for only
three thumb keys in use for each hand, but I think it's better to have
it there than needing a layer change. It used to be there so hopefully
the muscle memory will be restored quickly.