Replace Halmak with Hands Down layout
https://sites.google.com/alanreiser.com/handsdown

I don't suppose there is going to be much difference between the two,
but Hands Down is specifically created for split ortho, so it's probably
a little bit better.

I am not going with any of the variations that put keys on the thumb, I
think I've got enough going on with the thumbs to be getting on with.
The only change I've made is to keep the punctuation the same as Qwerty,
seems pretty pointless to shift that around.

One shot shift will not work with HD in this commit.