Remove one-shot shifts
Im not sure Ive ever really got the hang of it fully for one thing,
but more importantly I am about to start using a Helidox which only
has 3 thumbs, so something has to give. The plan is to move the
`enter` key right one place on the right thumb. Before I do that
I need to get used to not using the OSS, so lets do that now