Create Kyria keyboard profile
The layout has been extracted to jakeprime user space so that it can be
shared with multiple keyboards. As ever we are restricting to the
smallest keyboard in use, which is currently the Aurora Sweep which is
3x5 with 2 thumbs per side.

All the keys have been defined, but we are still missing screens, lights
and rotary encoders.