Separate out external LED strip
Previously the 8 external LEDs were mirrors of the onboard ones - they used the
same input. This works fine, but in the case where I want to turn off the
external ones without any soldering it's better to separate them, and be able to
control them separately.

This will still require a re-flashing, but that's really quick to do.

Note that the led map is not (as I had thought) the order the leds appear in
left to right, but rather then position on a horizontal range that each on in
order appears.

A couple of notes. This has stopped the screens working. I *think* this is
because the extra (thick) wire required to solder this is sapping current. I've
ordered some thinner AWG 30 wire in the hope that redoing that will help.

Also I've had to lower the max brightness to 220, or the keyboard crashed very
quickly.