Enable # and £ symbols using unicode on linux
These were not working as they were on Linux and I don't think there's any way
to enter a £ symbol from a US keyboard anyway. We could change the keyboard to a
UK one, but that would be quite the faff, the Mac keyboard basically is a US
layout other than £.

Instead we detect the OS (a nice surprise that this is possible with keyboard
firmware!) and use a macro to sent the unicode on Linux.